requirements.txt
tgcrypto
pyrogram==1.4.16
py-tgcalls==0.9.0
youtube-search-python
yt-dlp==2022.4.8
ffmpeg-python
aiohttp
aiofiles
asyncio
pytube
wget
python-dotenv
search-engine-parser
youtube-search
speedtest-cli
cryptography
dnspython
future
hurry.filesize
motor
Pillow
psutil
pykeyboard
requests
speedtest-cli
uvloop
telegraph
lyricsgenius
hachoir
rich
heroku3
gitpython
spotipy
pybase64
pyyaml
